+++ R/API_core.R
@@ -39,8 +39,11 @@
 #' @noRd
 
 .ddg.init <- function(prov.dir = NULL, overwrite = TRUE, save.debug = FALSE) {
+  preloaded = loadedNamespaces()
+  .ddg.set("ddg.preloaded.libraries", preloaded)
   
   # Initialize tables
+  #print ("Initiailzing tables")
   .ddg.init.tables()
 
   # Set up for console mode
@@ -50,22 +53,27 @@
   }
 
   # Get R script path
+  #print ("Getting R script path")
   r.script.path <- .ddg.r.script.path()
 
   # Set path for provenance graph
+  #print ("Setting provenance path")
   .ddg.set.path (prov.dir, r.script.path, overwrite)
   
   # Save value of save.debug
   .ddg.set("ddg.save.debug", save.debug)
 
   # Remove files from DDG directory
+  #print ("Removing ddg files")
   .ddg.flush.ddg()
 
   # Create DDG directories
+  #print ("Creating directories")
   .ddg.init.environ()
 
   # Script mode: adjust & store R script path
   if (.ddg.script.mode()) {
+    #print ("Setting up script mode")
 
     # RMarkdown script
     if (tools::file_ext(r.script.path) == "Rmd") {
@@ -88,6 +96,7 @@
   # completes execution and build the corresponding portions of the 
   # provenance graph.
   } else {
+    #print ("Setting up console mode")
     .ddg.store.console.info ()
     .ddg.set("ddg.markdown.output", NULL)
     .ddg.set("ddg.console.commands", vector())
@@ -108,12 +117,6 @@
     .ddg.set ("ddg.taskCallBack.id", addTaskCallback(.ddg.trace.task))
   }
 
-  # Store time when script begins execution.
-  .ddg.set("ddg.start.time", .ddg.timestamp())
-  
-  # Initialize the I/O tracing code
-  .ddg.init.iotrace ()
-  
   # Mark graph as initilized.
   .ddg.set("ddg.initialized", TRUE)
   
@@ -127,12 +130,27 @@
   
   # Initialize the table used to track use of non-locals
   # within functions
+  #print ("init func def table")
   .ddg.init.function.def.table ()
   
   # A named list, where the name is a variable of type environment
   # Associated with each environment name is a vector of the variables
   # in that environment. 
   .ddg.set ("ddg.envList", list())
+  
+  # Records the libraries loaded by the script itself
+  .ddg.set ("ddg.script.libraries", vector())
+  
+  # Store time when script begins execution.
+  .ddg.set("ddg.start.time", .ddg.timestamp())
+  
+  # Initialize the I/O tracing code
+  #print ("Init io tracing")
+  .ddg.init.iotrace ()
+  
+  
+  
+  #print ("returning from .ddg.init")
   
   invisible()
 }
@@ -262,6 +280,7 @@
 #' @noRd
 
 .ddg.quit <- function(save.debug = FALSE) {
+  #print ("In .ddg.quit")
   if (!.ddg.is.init()) return(invisible())
   
   # If running from the console create a Console finish node.
@@ -281,6 +300,7 @@
   
   # If there are any connections still open when the script ends,
   # create nodes and edges for them.
+  #print ("Closing connections")
   tryCatch(.ddg.create.file.nodes.for.open.connections (),
       error = function(e) {
         if (.ddg.debug.lib()) {
@@ -289,12 +309,14 @@
       })
   
   # If there is a display device open, grab what is on the display
+  #print ("Closing display device")
   if (length(grDevices::dev.list()) >= 1) {
     tryCatch (.ddg.capture.graphics(called.from.save = TRUE),
         error = function (e) print(e))
   }
   
   # If we ran an RMarkdown script, save a copy of the formatted output
+  #print ("Rmarkdown stuff")
   rmarkdown.output <- .ddg.get("ddg.markdown.output")
   if (! is.null (rmarkdown.output)) {
     if (file.exists (rmarkdown.output)) {
@@ -303,18 +325,21 @@
   }
   
   # Turn off the I/O tracing and console tracing.
+  #print ("Stopping I/O tracing.")
   .ddg.stop.iotracing()
   if (.ddg.is.set ("ddg.taskCallBack.id")) {
     removeTaskCallback (.ddg.get ("ddg.taskCallBack.id"))
   }
   
   # Delete temporary files.
+  #print ("Deleting temporary files")
   .ddg.delete.temp()
   
   # Mark graph as not initialized.
   .ddg.set("ddg.initialized", FALSE)
 
   # Save prov.json to file.
+  #print ("Writing json")
   .ddg.json.write()
   if (interactive()) print(paste("Saving prov.json in ", .ddg.path(), sep=""))
   
@@ -325,7 +350,7 @@
    
   # Set script mode to FALSE
   .ddg.set("ddg.script.mode", FALSE)
-
+  
   invisible()
 }
 
@@ -437,7 +462,8 @@
       else enc <- encoding
       if (length(enc) > 1L) {
         encoding <- NA
-        owarn <- options(warn = 2)
+        # In source function, but CRAN says not to change user's options
+        #owarn <- options(warn = 2)
         for (e in enc) {
           if (is.na(e)) 
             next
@@ -450,7 +476,8 @@
             break
           }
         }
-        options(owarn)
+        # Reverts to user's options.  Removed for CRAN
+        #options(owarn)
       }
       if (is.na(encoding)) 
         stop("unable to find a plausible encoding")